pyproject.toml
[project]
name = "project-8-modul"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "argon2-cffi>=25.1.0",
    "django>=5.2.7",
    "django-ckeditor-5>=0.2.18",
    "django-debug-toolbar>=6.0.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "flake8>=7.3.0",
    "flake8-bugbear>=24.12.12",
    "flake8-docstrings>=1.7.0",
    "flake8-import-order>=0.19.2",
    "isort>=6.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]
